トップページから入力フォームへの遷移を実装
