+++ app/helpers/bamboo_helper.rb
@@ -1,0 +1,2 @@
+module BambooHelper
+end
