+++ app/views/bamboo/new.html.erb
@@ -1,3 +1,1 @@
-<h4>思いの竹をつづる</h4>
-
 <%= render "form", bamboo: @bamboo %>
