--- app/views/home/_form.html.erb
@@ -1,16 +1,0 @@
-<%= form_with model: bamboo, url: home_create_path do |f| %>
-  <div>
-    <%= f.label :content, "内容" %><br>
-    <%= f.text_area :content %>
-
-    <% if bamboo.errors[:content].any? %>
-      <div style="color: red;">
-        <%= bamboo.errors[:content].join(", ") %>
-      </div>
-    <% end %>
-  </div>
-
-  <div>
-    <%= f.submit "送信" %>
-  </div>
-<% end %>
