+++ app/views/home/index.html.erb
@@ -1,2 +1,1 @@
-<h2>トップページ（仮）</h2>
-<p>ここに後で本番の内容を入れます。</p>
+<%= link_to "思いのたけをつづる", new_bamboo_path, class: "btn btn-primary" %>
