+++ config/routes.rb
@@ -1,6 +1,5 @@
 Rails.application.routes.draw do
   root "home#index"
 
-  get  "home/new",    to: "home#new"
-  post "home/create", to: "home#create"
+  resources :bamboo, only: [:new, :create]
 end
